permission_pasta/requirements.txt
boto3>=1.26.0
botocore>=1.29.0
requests>=2.28.0
python-dotenv>=0.21.0
PyGithub>=1.57.0
pyyaml>=6.0
regex>=2022.10.31
openai>=0.27.0  # Optional: For LLM-enhanced justification analysis 